% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shiny.R
\name{barchart}
\alias{barchart}
\title{Function to run an interactive Shiny-app to explore differences between
FCL and HS datasets.}
\usage{
barchart()
}
\description{
Function to run an interactive Shiny-app to explore differences between
FCL and HS datasets.
}

