% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{convertfao3}
\alias{convertfao3}
\title{Extract particular datasets from FAOSTAT3 data.}
\usage{
convertfao3(data, area, year, partner, flow = "all")
}
\arguments{
\item{area}{Numeric. Codes of reporter countries from FAO classification.}

\item{year}{Numeric. Reported year.}

\item{partner}{Numeric. Codes of partner countries from FAO classification.
All if omitted.}

\item{data.}{Data frame to take original data from. It should be downloaded
from faostat3.fao.org.}

\item{flow.}{Numeric. 1 - Import, 2 - Export. Both if omitted.}
}
\description{
Extract particular datasets from FAOSTAT3 data.
}

