% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{convertunctapi}
\alias{convertunctapi}
\title{Convert dataset gathered from UNCT API to comparable form.}
\usage{
convertunctapi(data, area, year, partner, flow = "all", weightonly = T)
}
\arguments{
\item{area}{Numeric. Codes of reporter countries from UNCT classification.
All if omitted.}

\item{year}{Numeric. Reported year. All if omitted.}

\item{partner}{Numeric. Codes of partner countries from UNCT classification.
All if omitted.}

\item{data.}{Data frame returned from call getunct() with compact = F}

\item{flow.}{Numeric. 1 - Import, 2 - Export. Both if omitted.}

\item{weightonly.}{Drop all except measured in kg. TRUE by default.}
}
\description{
Convert dataset gathered from UNCT API to comparable form.
}

