% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{faoarea}
\alias{faoarea}
\title{Convert UN ComTrade country codeinto corresponding FAO area code.}
\usage{
faoarea(unct)
}
\description{
Convert UN ComTrade country codeinto corresponding FAO area code.
}

