% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{faoareanames}
\alias{faoareanames}
\title{FAO areas names}
\format{A data frame with 293 observatins of 4 variables.
\describe{
  \item{fao}{FAO area code, numeric.}
  \item{iso2}{ISO2 area code, double character.}
  \item{name}{short area name, charecter.}
  \item{longname}{full area name, charecter.}
}}
\description{
FAO areas names
}

