% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/faostatdb.R
\name{faostatsql}
\alias{faostatsql}
\title{Sends given SQL-request to FAOSTAT Oracle DB
and returns resulting data.frame}
\usage{
faostatsql(sql, host = "lprdbwo1", port = 3310, db = "fstp",
  user = "demo", pass = "demo", classpath = "./ojdbc14.jar")
}
\description{
Sends given SQL-request to FAOSTAT Oracle DB
and returns resulting data.frame
}

