% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{fcl}
\alias{fcl}
\title{FCL items description.}
\format{A data frame with 776 observations of 6 variables.
\describe{
  \item{fcl}{FCL-code of item, numeric.}

}}
\description{
FCL items description.
}

