% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getfao}
\alias{getfao}
\title{Extract particular datasets from FAOSTAT3 data.}
\usage{
getfao(area, year, partner, flow = "all", compact = T, desc = F, data,
  debug = F)
}
\arguments{
\item{area}{Numeric. Codes of reporter countries from FAO classification.}

\item{year}{Numeric. Reported year.}

\item{partner}{Numeric. Codes of partner countries from FAO classification.
All if omitted.}

\item{flow.}{Numeric. 1 - Import, 2 - Export. Both if omitted.}

\item{compact.}{Logical. Should column names be standardized. TRUE by default.}

\item{desc.}{Logical. Should columns with descriptions be included.
FALSE by default.}

\item{data.}{Data frame to take original data from.}

\item{debug.}{Logical.}
}
\description{
Extract particular datasets from FAOSTAT3 data.
}

