% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/faostat3.R
\name{getfenix}
\alias{getfenix}
\title{Function retrieves data from FAOSTAT3 FENIX API}
\usage{
getfenix(list1Codes = NULL, list2Codes = NULL, list3Codes = NULL,
  list4Codes = NULL, list5Codes = NULL, list6Codes = NULL,
  list7Codes = NULL, nullValues = FALSE, thousand = "", decimal = ".",
  decPlaces = 2, datasource = "faostatdb", domainCode = "TM",
  lang = "E", url = "http://faostat3.fao.org/wds/rest/procedures/data")
}
\description{
JS source: https://gist.github.com/Kalimaha/80da85e6d7cbdbf92557
}

