% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/madb.R
\name{getmadb}
\alias{getmadb}
\title{Function to retrieve descriptions of country-specific
Harmonized System commodity codes from Market Access
Database of the European Commission.}
\usage{
getmadb(country, hs)
}
\arguments{
\item{country}{ISO3166 country code.}

\item{hs}{HS numeric code of length 4 or 6.}
}
\value{
Dataframe with columns country, hs code and description.
}
\description{
Function to retrieve descriptions of country-specific
Harmonized System commodity codes from Market Access
Database of the European Commission.
}

