% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/madb.R
\name{getmadbcountries}
\alias{getmadbcountries}
\title{Function to retrieve list of countries,
available at Market Access DB}
\usage{
getmadbcountries(url = "http://madb.europa.eu/madb/euTariffs.htm")
}
\description{
Function to retrieve list of countries,
available at Market Access DB
}

