% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/faostatdb.R
\name{gettfsource}
\alias{gettfsource}
\title{Returns data from FAOSTAT.TF_SOURCE_*}
\usage{
gettfsource(reporter = NULL, year = NULL, partner = NULL, flow = NULL,
  hs = NULL)
}
\description{
Returns data from FAOSTAT.TF_SOURCE_*
}

