% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/faostatdb.R
\name{gettfvalid}
\alias{gettfvalid}
\title{Returns data from FAOSTAT.TF_VALID}
\usage{
gettfvalid(reporter = NULL, year = NULL, partner = NULL, flow = NULL,
  fcl = NULL, ...)
}
\description{
Returns data from FAOSTAT.TF_VALID
}

