% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unsdapi.R
\name{getunct}
\alias{getunct}
\title{Get trading data via UNSD API.}
\usage{
getunct(area, year, partner = "all", flow = "all", code = "AG6",
  compact = T, desc = F, debug = F)
}
\description{
Get trading data via UNSD API.
}

