% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/growforest.R
\name{growforest}
\alias{growforest}
\title{Wrapper for CloudForest GO-lang RandomForest solution}
\usage{
growforest(data, sample, target, nTrees = 100, nSamples = 0, nCores = 1,
  progress = F, selftest = F, rm.unused = T, path = "/home/sas/go/bin/",
  workdir = "tmp")
}
\description{
Wrapper for CloudForest GO-lang RandomForest solution
}

