% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hsfclmap}
\alias{hsfclmap}
\title{FCL-HS2007 map.}
\format{A data frame with 1292 observations of  13 variables.}
\description{
Official correspondence map between FCL and HS2007 items.
}

