% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{joincompared}
\alias{joincompared}
\title{Function to join dataframes with FAO and UN CT data.}
\usage{
joincompared(faodata, ctdata)
}
\arguments{
\item{faodata.}{Data frame returned by convertfao3()}

\item{ctdata.}{Data frame returned by convertunctapi()}
}
\description{
Function to join dataframes with FAO and UN CT data.
}

