% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graph.R
\name{linksgroups}
\alias{linksgroups}
\title{Detect isolated groups of links between HS and FCL classification from
raw data.}
\usage{
linksgroups(hs, fcl)
}
\description{
Detect isolated groups of links between HS and FCL classification from
raw data.
}

