% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sws.R
\name{ping}
\alias{ping}
\title{Function ping send one ping-request to a server and returns TRUE
if gets a response.}
\source{
\url(https://github.com/mkao006/r_style_fao/blob/master/functions/sws_query.r)
}
\usage{
ping(x, stderr = FALSE, stdout = FALSE, ...)
}
\description{
Function ping send one ping-request to a server and returns TRUE
if gets a response.
}

