% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sws.R
\name{startjnc}
\alias{startjnc}
\title{Run jnc script to establish VPN-connection with FAO network.}
\usage{
startjnc(path = "/usr/local/bin/", profile = "fao")
}
\description{
Juniper VPN software has to be installed.
}

