% Generated by roxygen2 (4.0.2): do not edit by hand
\name{unctarea}
\alias{unctarea}
\title{Convert FAO area code into corresponding UN ComTrade country code.}
\usage{
unctarea(fao)
}
\description{
Convert FAO area code into corresponding UN ComTrade country code.
}
\section{Waring}{

It doesn't aggregate China's parts.
}

