% Generated by roxygen2 (4.0.2): do not edit by hand
\name{unctfaoareasmap}
\alias{unctfaoareasmap}
\title{FAO - UN Comtrade areas correspondence map.}
\format{A data frame with 241 observations of 2 variables.
\describe{
  \item{unct}{UN Comtrade area code, numeric.}
  \item{fao}{FAO area code, numeric.}
}}
\description{
FAO - UN Comtrade areas correspondence map.
}

