% Generated by roxygen2 (4.0.2): do not edit by hand
\name{uncttofao}
\alias{uncttofao}
\title{Convert data extracted from UN Comtrade API to FAOSTAT format.}
\usage{
uncttofao(data, long = T, faonames = T, quantity = F, netweight = F)
}
\description{
Convert data extracted from UN Comtrade API to FAOSTAT format.
}

