% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{unctarea}
\alias{unctarea}
\title{Convert FAO area code into corresponding UN ComTrade country code.}
\usage{
unctarea(fao)
}
\description{
Convert FAO area code into corresponding UN ComTrade country code.
}
\section{Waring}{

It doesn't aggregate China's parts.
}

