% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unsdapi.R
\name{uncttofao}
\alias{uncttofao}
\title{Convert data extracted from UN Comtrade API to FAOSTAT format.}
\usage{
uncttofao(data, long = T, faonames = T, quantity = F, netweight = F)
}
\description{
Convert data extracted from UN Comtrade API to FAOSTAT format.
}

